git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/branches/releases/v1.0@1490 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-tools/gendef/src/gendef.c
@@ -248,7 +248,8 @@ load_pep(void)
   fclose (fp);
   if (!gDta)
     {
-      fprintf (stderr, "*** [%s] unable to allocate %Iu bytes\n" ,fninput ,gDta_size);
+      fprintf (stderr, "*** [%s] unable to allocate %Iu bytes\n" ,fninput,
+	       (unsigned long) gDta_size);
       return 0;
     }
   gMZDta = (PIMAGE_DOS_HEADER) gDta;
